real alias IDs
this according to the plan I formulated earlier today. Let's create
random IDs and at the first sign of a conflict, extend the space from
which IDs are generated. This should keep alias names comparably short
while still providing reasonable randomness.

You will also notice the updatekey. We will require this to be set if an
existing alias should be updated or deleted using PUT or DELETE.