added configuration module
we have to use synchronized reading here - otherwise the code just gets
ever so much harder to read. On the positive node, modules are only
loaded on first require(), so this is basically nothing but startup
overhead.

Also, the merging of the config file with the defaults is really ugly,
but as the port requires special treatment (parseInt) and there's not
that much more left to do, the alternative of using for...in with
hasOwnProperty doesn't make sense (because it would just catch one
setting that's not an int)