hook into "MAIL FROM" and "RCPT TO"
update to node-smtp trunk which provides the needed facilities.
Afterwards, write handlers for mail from and rcpt to and pass the data
received directly to the upstram server.

Also, immediately return any errors reported by upstream and close the
connection.

Also, note the replacement of the to address in the rcpt_to handler.
This is where we are going to hook into our actual backend code to
retrieve the real to-address