only connect to upstream at RCPT time
we are providing autodestructing service, so it's likely that the common
case for any mail delivery will be the delivery of a message to an
expired alias.

This means that in the majority of the time, we don't have to bother the
upstream server but can reject at RCPT time.

This is why I think it's way better to only connect after our (currently
non existant) RCPT TO checks have passed.

Of course this might yield strange behaviour for any connecting server
if upstream has a problem as we would report, say, an incorrect FROM
only at RCPT TO time.