check for alias validity
Alias gained a use() method that calls back with either true or false if
the alias could be used. In doing so, it checks for time based validity
and decrements the usage counter.

findById() gained a 'use' parameter to allow choice between finding the
alias just for looking at it or finding it with the intention to use it.

The SMTP server obviously does the latter