drop express for paperboy
see blog post for reasoning, but here's the executive summary:

- express did too much (too much abstraction, too many dependencies, too
  many features. none needed) and too little (little help for webservice
  like requests)
- express had a strange issue that forced node not to cleanly close a
  connection if the request body was null (for 304: not modified)

The nice thing about paperboy is that it doesnt try to be rails, but
it's just a plug-in solution to plug static file serving as a feature of
the webserver you are creating.

This is SO MUCH nicer.

And the rewrite isn't that big either